Rendering and scrolling working
